requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
comm==0.2.2
datasets==2.20.0
debugpy==1.8.2
decorator==5.1.1
dill==0.3.8
distro==1.9.0
executing==2.0.1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.5.0
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==5.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tokenizers==0.19.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers.git@1218e439b5fe05423693407653a1fb064263fef4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
easyocr~=1.7.1